clean up, added export functionality, worked on print messages

+++ src/conote/modules/controller.py
@@ -1,18 +1,23 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python3      
 
 import os
 from pathlib import Path
+from sys import path_importer_cache
 from rich.console import Console
 from rich.table import Table
 from rich import box
 from rich.markdown import Markdown
 from rich.panel import Panel
 from conote.modules import MarkDownInit, NoteDataBase
+from conote.modules.tools import PATH, stamps, messages, tables
 
+NOTE_DIR = Path.home() / "Documents/CNotes"
 
 class NoteCommandController(NoteDataBase, MarkDownInit):
     
     def __init__(self) -> None:
+
+        self.note_mkdir
         NoteDataBase.__init__(self)
         MarkDownInit.__init__(self)
 
@@ -21,37 +26,35 @@ class NoteCommandController(NoteDataBase, MarkDownInit):
         self.console = Console()
 
     
-    def add(self, note):
+    def add(self, note) -> None:
         
         self.create_note(note)
         self.insert()
 
 
+    def ls(self) -> None:
 
-    def ls(self):
+        path = self.select()
 
-        table = Table(title='Notes List', box=box.HORIZONTALS, caption='END OF LIST', show_lines=False, header_style='red')
+        if path:
 
-        table.add_column("ID", justify="right", style="red", no_wrap=True)
-        table.add_column("NAME", justify="right", style="cyan", no_wrap=True)
-        table.add_column("CREATED", justify="right", style="cyan", no_wrap=True)
-                
-        for record in self.select():         
+            for record in self.select():         
             
-            table.add_row(f'{record[0]}', f'{record[1]}', f'{record[2]}')
+                tables.add_row(f'{record[0]}', f'{record[1]}', f'{record[2]}')
 
-        return self.console.print(table)
+        return self.console.print(tables)
 
-    def rm(self, id: int):
+
+
+    def rm(self, id: int) -> None:
 
         path = self.is_exists(id)
 
-        if path:
+        if path and self.drop(id) is not None:
             
            os.remove(path)
-           self.drop(id)
                 
-           print('file sucsessfully deleted')
+           messages.info.print('file sucsessfully deleted')
 
         else:
             
@@ -67,8 +70,11 @@ class NoteCommandController(NoteDataBase, MarkDownInit):
  
             return self.console.print(Panel(md, title=f'{path}', box=box.HORIZONTALS, expand=True))
 
+        else:
+            print('No note files for showing')
+
     
-    def is_exists(self, id: int):
+    def is_exists(self, id: int) -> str | None:
         
         record: tuple = self.select(id=id)
 
@@ -76,15 +82,30 @@ class NoteCommandController(NoteDataBase, MarkDownInit):
             
             file = "".join([record[1], '.md',])
         
-            file_path = os.path.join(self.dbase_dir, file)
+            file_path = os.path.join(PATH, file)
             if os.path.exists(file_path):
                 return file_path
         else:
             return None
         
-    def export(self):
-        pass
 
+    def export(self, id: int, pattern: str) -> None:
+
+        path = self.is_exists(id)
+        
+        if path:
+            file_name = path.split('.')[0]
+
+            raw_md = self.read_file(path)
+
+            self.exporter(path=file_name, raw_file=raw_md, pattern=pattern)
+
+        else:
+            print('file note exists')
+
+    @property
+    def note_mkdir(self):
+        os.makedirs(PATH, exist_ok=True)
 
 if __name__ == '__main__':
     note = NoteCommandController()
